rename components and switch utils to const
